requirements.txt
flask
flask-cors
python-dotenv
google-generative-ai